requirements.txt
Flask
python-decouple
langchain==0.1.14
openai
pydantic
tiktoken
pymongo
python-dotenv